+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add `sha1Hash` method
 - Add `sha256Hash` method
 - Add `encodeBase64` method
+- Add `decodeBase64` method
 
 ## 0.0.4 (2024-10-02)